Change approach on Exercice 2
Insted of creating new variables (allOptions, toDelete) which
contained a list of nodes and parsing it 'manually' to find the proper
node to delete I decided to take advantage of the <select> element's
method .remove(index) and property .selectedIndex.

*Since there was no straightforward way to benefit from the array
method,filter, because the only solution that I came up with
supposed to: use the initial NodeList to create a new array then filter
it, then create a new Node list and replace it, i gave up on that.